code review section re-design
